# swagger.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/hr/leaves":  {
                                         "post":  {
                                                      "operationId":  "LeavesController_create",
                                                      "parameters":  [

                                                                     ],
                                                      "requestBody":  {
                                                                          "required":  true,
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  {
                                                                                                                                  "$ref":  "#/components/schemas/CreateLeaveDto"
                                                                                                                              }
                                                                                                               }
                                                                                      }
                                                                      },
                                                      "responses":  {
                                                                        "201":  {
                                                                                    "description":  ""
                                                                                }
                                                                    },
                                                      "tags":  [
                                                                   "Leaves"
                                                               ]
                                                  },
                                         "get":  {
                                                     "operationId":  "LeavesController_findAll",
                                                     "parameters":  [
                                                                        {
                                                                            "name":  "employeeId",
                                                                            "required":  true,
                                                                            "in":  "query",
                                                                            "schema":  {
                                                                                           "type":  "string"
                                                                                       }
                                                                        }
                                                                    ],
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  ""
                                                                               }
                                                                   },
                                                     "tags":  [
                                                                  "Leaves"
                                                              ]
                                                 }
                                     },
                  "/api/hr/leaves/{id}/review":  {
                                                     "patch":  {
                                                                   "operationId":  "LeavesController_review",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "required":  true,
                                                                                          "in":  "path",
                                                                                          "schema":  {
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "required":  true,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/ReviewLeaveDto"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  ""
                                                                                             }
                                                                                 },
                                                                   "tags":  [
                                                                                "Leaves"
                                                                            ]
                                                               }
                                                 },
                  "/api/hr/attendance/check-in":  {
                                                      "post":  {
                                                                   "operationId":  "AttendanceController_checkIn",
                                                                   "parameters":  [

                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "required":  true,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/PunchDto"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                   "responses":  {
                                                                                     "201":  {
                                                                                                 "description":  ""
                                                                                             }
                                                                                 },
                                                                   "tags":  [
                                                                                "Attendance"
                                                                            ]
                                                               }
                                                  },
                  "/api/hr/attendance/check-out":  {
                                                       "post":  {
                                                                    "operationId":  "AttendanceController_checkOut",
                                                                    "parameters":  [

                                                                                   ],
                                                                    "requestBody":  {
                                                                                        "required":  true,
                                                                                        "content":  {
                                                                                                        "application/json":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/PunchDto"
                                                                                                                                            }
                                                                                                                             }
                                                                                                    }
                                                                                    },
                                                                    "responses":  {
                                                                                      "201":  {
                                                                                                  "description":  ""
                                                                                              }
                                                                                  },
                                                                    "tags":  [
                                                                                 "Attendance"
                                                                             ]
                                                                }
                                                   },
...
                  "/api/hr/attendance/{employeeId}/summary":  {
                                                                  "get":  {
                                                                              "operationId":  "AttendanceController_getSummary",
                                                                              "parameters":  [
                                                                                                 {
                                                                                                     "name":  "employeeId",
                                                                                                     "required":  true,
                                                                                                     "in":  "path",
                                                                                                     "schema":  {
                                                                                                                    "type":  "string"
                                                                                                                }
                                                                                                 },
                                                                                                 {
                                                                                                     "name":  "month",
                                                                                                     "required":  true,
                                                                                                     "in":  "query",
                                                                                                     "schema":  {
                                                                                                                    "type":  "string"
                                                                                                                }
                                                                                                 },
                                                                                                 {
                                                                                                     "name":  "year",
                                                                                                     "required":  true,
                                                                                                     "in":  "query",
                                                                                                     "schema":  {
                                                                                                                    "type":  "string"
                                                                                                                }
                                                                                                 }
                                                                                             ],
                                                                              "responses":  {
                                                                                                "200":  {
                                                                                                            "description":  ""
                                                                                                        }
                                                                                            },
                                                                              "tags":  [
                                                                                           "Attendance"
                                                                                       ]
                                                                          }
                                                              },
...
              },
    "info":  {
                 "title":  "Trifusion-Dynamics Auth Service",
                 "description":  "The API description for Auth Service",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
                       "schemas":  {
...
                                       "CreateLeaveDto":  {
                                                              "type":  "object",
                                                              "properties":  {

                                                                             }
                                                          },
                                       "ReviewLeaveDto":  {
                                                              "type":  "object",
                                                              "properties":  {

                                                                             }
                                                          },
                                       "PunchDto":  {
                                                        "type":  "object",
                                                        "properties":  {

                                                                       }
                                                    },
...
                                   }
                   }
}